recvfrom

+++ cpkg/lib.c
@@ -19,7 +19,7 @@ unsigned int local_network_mask[4] = { 10, 127, 4268, 43200 };
 
 typedef int (*Connect_callback)(int, const struct sockaddr*, socklen_t);
 typedef ssize_t (*Sendto_callback)(int, const void*, size_t, int, const struct sockaddr*, socklen_t);
-typedef int (*Recvfrom_callback)(int, void*, size_t, int, struct sockaddr*, socklen_t*);
+typedef ssize_t (*Recvfrom_callback)(int, void*, size_t, int, struct sockaddr*, socklen_t*);
 
 Connect_callback __real_connect = NULL;
 Sendto_callback __real_sendto = NULL;
@@ -49,7 +49,7 @@ ssize_t real_sendto(int s, const void *msg, size_t len, int flags, const struct 
     __real_sendto(s, msg, len, flags, to, tolen);
 }
 
-int real_recvfrom(int s, void *buf, size_t len, int flags, struct sockaddr *from, socklen_t *fromlen){
+ssize_t real_recvfrom(int s, void *buf, size_t len, int flags, struct sockaddr *from, socklen_t *fromlen){
     if (__real_recvfrom == NULL) {
         void *h_dl = RTLD_NEXT;
         if (h_dl == NULL) {
@@ -214,7 +214,7 @@ SO_EXPORT ssize_t sendto(int s, const void *msg, size_t len, int flags, const st
     return __real_sendto(s, msg, len, flags, to, tolen);
 }
 
-SO_EXPORT int recvfrom(int s, void *buf, size_t len, int flags, struct sockaddr *from, socklen_t *fromlen){
+SO_EXPORT ssize_t recvfrom(int s, void *buf, size_t len, int flags, struct sockaddr *from, socklen_t *fromlen){
     if(socket_type(s) == SOCK_DGRAM){
         errno = ECONNREFUSED;
         return -1;
